Remove unneeded cefclient.h and cefclient.cpp files
