+++ Atom-Linux/atom.h
